requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
click==8.0.3
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
PyJWT==1.7.1
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
